Re-work the class system
Add a separate root class for classes without dynamic size so that we
can abort when the user mixes them up.

Auto-call the SETUP method when creating the object.

In "voc: X class: A … class: B" A and B should be in X. That didn't
always work.

Drop old MM version history.

Move class-sized.fs into class.fs, no need for that to be separate.

Don't use SIZE to configure variable-sized objects.